Extract class to model JMSMetaData to increase cohesion among class methods.
